% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinityMatrix.R
\name{affinityMatrix}
\alias{affinityMatrix}
\title{To calculate the similarity matrix}
\usage{
affinityMatrix(Diff, K = 20, sigma = 0.5)
}
\arguments{
\item{Diff}{A diff matrix}

\item{K}{The number of neighbors in consideration}

\item{sigma}{A parameter to determine the scale}
}
\value{
W The similarity matrix
}
\description{
calculate the affinity matrix from the diff matrix with 20 neighbors
}
\examples{

data_list <- simu_data_gen()
Diff <- dist2eu(Standard_Normalization(data_list[[1]]),Standard_Normalization(data_list[[1]]))
simi <- affinityMatrix(Diff,20,0.5)
}
\author{
Xiaoyao Yin
}
