% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cal_NMI.R
\name{Cal_NMI}
\alias{Cal_NMI}
\title{calculate the normalized mutual information.}
\usage{
Cal_NMI(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector as long as x}
}
\value{
A number between 0 and 1 indicating the normalized mutual information
}
\description{
calculate the normalized mutual information of two vectors x and y.
}
\examples{

x <- c(0.1,0.2,0.3,0.4)
y <- c(0.1,0.2,0.3,0.4)
NMI <- Cal_NMI(x,y)
}
\author{
Xiaoyao Yin
}
