% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2bin.R
\name{dist2bin}
\alias{dist2bin}
\title{Calculate the agreement-based measurement}
\usage{
dist2bin(X, C)
}
\arguments{
\item{X}{A sample-feature matrix with rows as samples and columns as features}

\item{C}{The same as X}
}
\value{
A matrix whose elements at (i,j) is the agreement-based measurement of two any pair-wise samples x_i and x_j
}
\description{
Calculate the agreement-based measurement of two any pair-wise samples x_i and x_j for binary variables
}
\examples{

data_list <- simu_data_gen()
X <- data_list[[1]]
C <- X
Diff <- dist2bin(X,C)
}
\author{
Xiaoyao Yin
}
