% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2chi.R
\name{dist2chi}
\alias{dist2chi}
\title{Calculate the chi-squared distance}
\usage{
dist2chi(X, C)
}
\arguments{
\item{X}{A sample-feature matrix with rows as samples and columns as features}

\item{C}{The same as X}
}
\value{
A matrix whose elements at (i,j) is the chi-squared distance of two any pair-wise samples x_i and x_j
}
\description{
Calculate the chi-squared distance of two any pair-wise samples x_i and x_j for discrete variables
}
\examples{

data_list <- simu_data_gen()
X <- data_list[[1]]
C <- X
Diff <- dist2chi(X,C)
}
\author{
Xiaoyao Yin
}
