% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2eu.R
\name{dist2eu}
\alias{dist2eu}
\title{Calculate the Euclidean distance}
\usage{
dist2eu(X, C)
}
\arguments{
\item{X}{A sample-feature matrix with rows as samples and columns as features}

\item{C}{The same as X}
}
\value{
A matrix whose elements at (i,j) is the Euclidean distance of two any pair-wise samples x_i and x_j
}
\description{
Calculate the Euclidean distance of two any pair-wise samples x_i and x_j for continuous variables
}
\examples{

data_list <- simu_data_gen()
X <- data_list[[1]]
C <- X
Diff <- dist2eu(X,C)
}
\author{
Xiaoyao Yin
}
