% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initialization}
\alias{initialization}
\title{initialize the sub-matrix Ci into alpha*Li by SVD}
\usage{
initialization(WL, k)
}
\arguments{
\item{WL}{A list of similarity matrices}

\item{k}{A parameter to specify the cluster number}
}
\value{
A list with N+2 elements, the former N as modality private sub-matrices, the Nth as the shared sub-matrix and the last one as 1
}
\description{
Li takes the first k columns of matrix d in SVD, while alpha is the mean of all the u of SVD result in each modality
}
\examples{

WL <- simu_data_gen()
new_WL_list <- initialize_WL(WL)
k <- 5
init_list <- initialization(new_WL_list,k)
}
\author{
Xiaoyao Yin
}
