% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_WL.R
\name{initialize_WL}
\alias{initialize_WL}
\title{Initialize from the similairty matrix list}
\usage{
initialize_WL(WL)
}
\arguments{
\item{WL}{A list of similarity matrices}
}
\value{
A list as long as WL with elements satisfying res[[i]]%*%t(res[[i]])=WL[[i]]
}
\description{
Factorize the each of the similairty matrix Si into Ci*t(Ci) by SVD
}
\examples{

WL <- simu_data_gen()
new_WL_list <- initialize_WL(WL)
}
\author{
Xiaoyao Yin
}
