% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M2SMJF.R
\name{M2SMJF}
\alias{M2SMJF}
\title{the main part for M2SMJF and clustering result}
\usage{
M2SMJF(WL, lambda = 0.25, theta = 10^-4, k)
}
\arguments{
\item{WL}{A list of similarity matrices}

\item{lambda}{A parameter to set the relative weight of the group sparsity constraints}

\item{theta}{A parameter to determine the convergence}

\item{k}{A parameter to specify the cluster number}
}
\value{
A list containing the clustering result
\item{sub_matrices}{a list containing all the sub-matrices}
\item{cluster_res}{the clustering result which is as long as the number of samples}
}
\description{
jointly factorize multiple matrices into a shared sub-matrix and multiple private sub-matrices
}
\examples{

WL <- simu_data_gen()
res <- M2SMJF(WL,0.25,10^-4,5)
}
\author{
Xiaoyao Yin
}
