% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_data_gen.R
\name{simu_data_gen}
\alias{simu_data_gen}
\title{Generate simulated data}
\usage{
simu_data_gen()
}
\value{
A list with two elements, which are the sample-feature matrices from different modality
}
\description{
A function to generate simulated data with two modularities and five clusters
}
\examples{

data_list <- simu_data_gen()
}
\author{
Xiaoyao Yin
}
