% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standard_Normalization.R
\name{Standard_Normalization}
\alias{Standard_Normalization}
\title{Normalize the input matrix by column}
\usage{
Standard_Normalization(x)
}
\arguments{
\item{x}{A sample-feature matrix with rows as samples and columns as features}
}
\value{
A sample-feature matrix with rows as samples and columns as features,each column of the matrix have mean 0 and standard deviation 1
}
\description{
Normalize each column of x to have mean 0 and standard deviation 1.
}
\examples{

data_list <- simu_data_gen()
x <- data_list[[1]]
data_matrix <- Standard_Normalization(x)
}
\author{
Xiaoyao Yin
}
